# manage-service/docs/swagger.yaml
definitions:
...
  entity.EvaluationCriteria:
    properties:
      code:
        type: string
      max_score:
        type: integer
      title:
        type: string
      type:
        description: BASE, ALTERNATIVE
        type: string
    type: object
  entity.ExpertEvaluation:
    properties:
      applicant_id:
        type: integer
      category:
        type: string
      comment:
        type: string
      created_at:
        type: string
      expert_id:
        type: string
      id:
        type: integer
      is_admin_override:
        type: boolean
      score:
        type: integer
      source_info:
        type: string
      status:
        description: DRAFT, COMPLETED
        type: string
      updated_at:
        type: string
      updated_by_id:
        type: string
    type: object
  entity.ExpertSlot:
    properties:
      created_at:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      slot_number:
        type: integer
      user_id:
        type: string
    type: object
  entity.Program:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      status:
        type: string
      title:
        type: string
      year:
        type: integer
    type: object
  entity.User:
    properties:
      email:
        type: string
      first_name:
        type: string
      id:
        type: string
      last_name:
        type: string
      role:
        type: string
    type: object
  handlers.assignSlotRequest:
    properties:
      role:
        type: string
      slot_number:
        type: integer
      user_id:
        type: string
    type: object
...
  handlers.expertEvaluationRequest:
    properties:
      complete:
        type: boolean
      expert_id:
        type: string
      scores:
        items:
          $ref: '#/definitions/entity.ExpertEvaluation'
        type: array
      user_id:
        type: string
      user_name:
        type: string
      user_role:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: API for managing applicants and documents
  title: Manage Service API
  version: "1.0"
paths:
...
  /v1/applicants/{id}/criteria:
    get:
      description: Возвращает список всех критериев экспертной оценки. ID абитуриента
        в пути используется для валидации.
      parameters:
      - description: ID абитуриента
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.EvaluationCriteria'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Получение критериев оценки
      tags:
      - experts
...
  /v1/applicants/{id}/evaluations:
    get:
      description: 'Возвращает список экспертных оценок по абитуриенту. Результат
        фильтруется по роли пользователя: эксперт видит только свои оценки, администратор
        — все.'
      parameters:
      - description: ID абитуриента
        in: path
        name: id
        required: true
        type: integer
      - description: ID текущего пользователя
        in: query
        name: user_id
        type: string
      - description: Роль текущего пользователя
        in: query
        name: user_role
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.ExpertEvaluation'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Список оценок экспертов
      tags:
      - experts
    put:
      consumes:
      - application/json
      description: Сохраняет или обновляет оценки эксперта по критериям для указанного
        абитуриента. Завершённую оценку изменить нельзя.
      parameters:
      - description: ID абитуриента
        in: path
        name: id
        required: true
        type: integer
      - description: Данные оценки
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.expertEvaluationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Сохранение экспертной оценки
      tags:
      - experts
...
  /v1/experts:
    get:
      description: Возвращает список всех пользователей с ролью эксперта.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.User'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Список экспертов
      tags:
      - experts
  /v1/experts/slots:
    get:
      description: Возвращает все слоты назначения экспертов с их статусами.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.ExpertSlot'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Список слотов экспертов
      tags:
      - experts
    post:
      consumes:
      - application/json
      description: Назначает пользователя на указанный слот эксперта. Доступно только
        администратору.
      parameters:
      - description: Данные назначения
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.assignSlotRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Назначение эксперта на слот
      tags:
      - experts
  /v1/programs:
    get:
      description: Получение списка доступных магистерских программ с пагинацией (по
        умолчанию limit=100)
      parameters:
      - description: Количество записей (по умолчанию 100)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.Program'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Список программ
      tags:
      - programs
  /v1/programs/{id}:
    get:
      description: Получение информации об образовательной программе по ID
      parameters:
      - description: ID программы
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Program'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Информация о программе
      tags:
      - programs
securityDefinitions:
  BearerAuth:
    description: 'Введите токен в формате: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"